about to start fine tuning bidding prices, dont wanna lose the progress

+++ commerce/auctions/models.py
@@ -9,8 +9,8 @@ class Listings(models.Model):
     title = models.CharField(max_length=128)
     description = models.TextField(max_length=500)
     image = models.URLField(blank=True)
-    listing_start_price = models.IntegerField()
-    current_price = models.IntegerField(default=0)
+    listing_start_price = models.DecimalField(decimal_places=2,max_digits=10)
+    current_price = models.DecimalField(default=0,decimal_places=2,max_digits=10)
     seller = models.ForeignKey(User, null=True, on_delete=models.CASCADE, related_name="seller")
     watchers = models.ManyToManyField(User, blank=True, related_name="current_watchers")
     active = True
@@ -30,7 +30,7 @@ class Listings(models.Model):
 class Bids(models.Model):
     item = models.ForeignKey(Listings, on_delete=models.CASCADE, null=True, related_name="bid_for_item")
     bidder = models.ForeignKey(User,on_delete=models.CASCADE, null=True, related_name="bid_made")
-    bid_price = models.IntegerField()
+    bid_price = models.DecimalField(decimal_places=2,max_digits=10)
     
     def __str__(self) -> str:
         return f'{self.item} bid'
@@ -42,7 +42,7 @@ class Bids(models.Model):
         self.item = item
 
     def get_bid_price(self):
-        return int(self.bid_price)
+        return self.bid_price
 
     def accept_bid(self, item):
         if (item.listing_start_price < self.bid_price) and (item.current_price<self.bid_price):
